refactor: 금융상품 비교 페이지 정리
